Clean up 'prompt user for info' modals
